+++ lib/widgets/twitterapp.dart
@@ -11,7 +11,7 @@ import '../models/user.dart';
 
 import 'package:intl/intl.dart';
 import '../models/credential.dart';
-
+import './twitimer.dart';
 
 class TwitterApp extends StatefulWidget {
   const TwitterApp({Key? key}) : super(key: key);
@@ -23,34 +23,7 @@ class TwitterApp extends StatefulWidget {
 class _TwitterAppState extends State<TwitterApp> {
   static final _dateFormatter = DateFormat('E MMM dd HH:mm:ss yyyy');
 
-  late Timer _timer;
-  late DateTime _time;
-
-
-  @override
-  void initState(){
-    _time = DateTime.utc(0,0,0,0,0,10);
-    _startTimer();
-    super.initState(); //ちゃんと継承している
-  }
-  void _startTimer() {
-    {
-      _timer = Timer.periodic(Duration(seconds: 10),
-              (Timer timer) => setState((){_time = _time.subtract(Duration(seconds: 10));
-          if(_time.isAtSameMomentAs(DateTime.utc(0,0,0,0,0,0))){
-            _timer.cancel();
-
-            Navigator.of(context).pushReplacementNamed('/timer');
-            //移動先のサービスをプッシュする。
-          }
-          })
-      );
-    }
-  }
-
   Future<List<Tweet>> getTimeline(credential) async {
-    print("getTimeline");
-
     final client = credential.userclient;
     final apiResponse = await client.get(
       Uri.parse(
@@ -72,8 +45,6 @@ class _TwitterAppState extends State<TwitterApp> {
         favorited: tweet['favorited'],
       );
       _timelines.add(t);
-      print(_timelines);
-
     }
     return Future.value(_timelines);
   }
@@ -83,7 +54,7 @@ class _TwitterAppState extends State<TwitterApp> {
     final credential = Provider.of<Credential>(context);
     return Scaffold(
       appBar: AppBar(
-        title: Text(DateFormat.ms().format(_time)),
+        title: twitimer(),
       ),
       body: Center(
         child:
